scripts/requirements.txt
langchain
python-frontmatter
markdown
beautifulsoup4
unstructured
supabase
sentence-transformers
langchain-community